Extracts TraceableAdapterEvent class to it's own file
